+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.6.0] - 2026-06-12
 
 ### Changed
 - Added support for 'method' parameter on Replacer 'add_rule'
@@ -211,7 +211,7 @@ ensure it's automatically sent in all API requests.
 ### Changed
 - Moved from the main `zaproxy` repository.
 
-[Unreleased]: https://github.com/zaproxy/zap-api-python/compare/0.5.0...HEAD
+[0.6.0]: https://github.com/zaproxy/zap-api-python/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/zaproxy/zap-api-python/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/zaproxy/zap-api-python/compare/0.3.2...0.4.0
 [0.3.2]: https://github.com/zaproxy/zap-api-python/compare/0.3.1...0.3.2